docker-crawler/crawler/requirements.txt
scrapy>=2.11.0
scrapy-playwright>=0.0.34
playwright>=1.40.0
psycopg2-binary>=2.9.9
requests>=2.31.0
beautifulsoup4>=4.12.0
lxml>=5.0.0
python-dateutil>=2.8.2
spacy>=3.7.0
python-whois>=0.8.0
# Anti-detection & Stealth
fake-useragent>=1.4.0
playwright-stealth>=1.0.6
# Proxy rotation
scrapy-rotating-proxies>=0.6.2
# Redis queue
redis>=5.0.0
# Retry & resilience
tenacity>=8.2.0
# Tor support
PySocks>=1.7.1
stem>=1.8.2
# Improved Extraction
trafilatura>=1.6.0
phonenumbers>=8.13.0
thefuzz>=0.20.0
# Compression support
brotli>=1.1.0
